refactor: a tooltip component to use mui props and merge into a single tsx file
BREAKING CHANGE: `message` props was replaced with `title`
