+++ src/components/Tooltip/Tooltip.tsx
@@ -1,7 +1,13 @@
+import { TooltipProps as MuiTootipProps } from '@material-ui/core/Tooltip';
+import { forwardRef } from 'react';
 import styled, { css } from 'styled-components';
 import MuiTooltip from '@material-ui/core/Tooltip';
-import { TooltipProps } from './TooltipContainer';
 import { defaultProps } from '../../theme';
+
+export type TooltipProps = Pick<
+  MuiTootipProps,
+  'id' | 'title' | 'placement' | 'arrow' | 'className' | 'children' | 'classes'
+>;
 
 const StyledTooltip = styled(({ theme, className, ...rest }) => (
   <MuiTooltip
@@ -11,7 +17,7 @@ const StyledTooltip = styled(({ theme, className, ...rest }) => (
 ))<TooltipProps>`
   ${({ theme: { tooltip } }) => css`
     .tooltip {
-      height: ${tooltip.height};
+      margin: ${tooltip.margin};
       padding: ${tooltip.padding};
       background-color: ${tooltip.backgroundColor};
       border-radius: ${tooltip.borderRadius};
@@ -26,24 +32,19 @@ const StyledTooltip = styled(({ theme, className, ...rest }) => (
 `;
 StyledTooltip.defaultProps = defaultProps;
 
-export const TooltipPresenter: React.FC<TooltipProps> = ({
-  message,
-  placement = 'bottom',
-  arrow = true,
-  children,
-  className,
-  ...rest
-}) => {
-  return (
-    <StyledTooltip
-      arrow={arrow}
-      title={message}
-      placement={placement}
-      {...rest}
-    >
+const Tooltip = forwardRef<HTMLDivElement, TooltipProps>(
+  (
+    { placement = 'bottom', arrow = true, children, className, ...rest },
+    ref
+  ) => (
+    <StyledTooltip ref={ref} arrow={arrow} placement={placement} {...rest}>
       <div style={{ display: 'inline-block' }} className={className}>
         {children}
       </div>
     </StyledTooltip>
-  );
-};
+  )
+);
+
+Tooltip.displayName = 'Tooltip';
+
+export { Tooltip };
